:recycle: refactor: rename useProgramStorage to better reflect usage and functionality
